+++ src/utils/string_utils.tsx
@@ -24,9 +24,9 @@ export function escapeString(str: string, quoteType: '"' | "'" = '"'): string {
 }
 
 export function validateDelimiter(delimiter: string): boolean {
-    // only allow special characters as delimiters 
+    // only allow special characters as delimiters. disallow unsafe characters { }, $ 
     if (!(/^[^a-zA-Z0-9\s]+$/.test(delimiter))) return false;
-    if (/[{}]/.test(delimiter)) return false;
+    if (/[{}$]/.test(delimiter)) return false;
     return true;
 }